+++ docente/modulisticaRichiestaApprova.php
@@ -15,9 +15,7 @@ require_once '../common/checkSession.php';
 // ruoloRichiesto('docente','dirigente','segreteria-docenti','segreteria-didattica');
 require_once '../common/connect.php';
 require_once '../common/header-common.php';
-
-function produciTabella() {
-}
+require_once '../docente/modulisticaProduciTabella.php';
 
 function notifica($to, $toName, $titolo, $stato, $messaggio) {
 	global $nomeCognomeDocente, $template;
@@ -34,8 +32,8 @@ function notifica($to, $toName, $titolo, $stato, $messaggio) {
 	$mail->Body .= "<p>Il docente ".$nomeCognomeDocente." ha inviato il modulo ".$template['nome'];
 
 	$mail->Body .= '<p>La richiesta '.$titolo.' &egrave; stata '.$stato.' in data '.date("d/m/Y").'</p>';
-
-	$mail->Body .= produciTabella();
+// nota: il messaggio in questo momento e' ancora vuoto perche' viene inserito in seguito, per cui lo lasciamo perdere per il momento
+//	$mail->Body .= produciTabella();
 
 	$mail->Body .= "</body></html>";
 	$mail->AltBody = 'Richiesta '.$titolo.': '.$stato.' in data '.date("d/m/Y");
@@ -83,34 +81,42 @@ if ($richiesta['data_approvazione'] != null) {
 }
 setlocale(LC_TIME, $oldLocale);
 
-$titolo = '[M-' . $richiesta_id . '] ' . $template['nome'] .' - ' . $nomeCognomeDocente;
+$titolo = 'M-' . $richiesta_id . ' ' . $template['nome'] .' - ' . $nomeCognomeDocente;
 
-// se e' gia' in stato di approvata, respinta o annullata non prende altri comandi
-if (! $approvata && ! $respinta && ! $annullata) {
+// controlla se e' stata richiesta l'esecuzione di un comando
+if(isset($_GET['comando'])) {
+	$comando = $_GET['comando'];
 
-	// se c'e' un comando, lo esegue
-	if(isset($_GET['comando'])) {
-		$comando = $_GET['comando'];
+	// controlla se eì stato passato un messaggio
+	if(isset($_GET['messaggio'])) {
+		$messaggio = $_GET['messaggio'];
+	} else {
+		$messaggio = "";
+	}	
 
+	// se va chiusa si chiude e basta
+	if ($comando == 'chiudi') {
+		dbExec('UPDATE modulistica_richiesta SET `chiusa` = 1, data_chiusura=now() WHERE id = '.$richiesta_id.';');
+		info("chiusa la richiesta id=$richiesta_id richiesta=$titolo");
+	}
+
+	// se e' gia' in stato di approvata, respinta o annullata non prende altri comandi a parte chiudi
+	else if (! $approvata && ! $respinta && ! $annullata) {
 		if ($comando == 'approva') {
-			dbExec("UPDATE modulistica_richiesta SET `approvata` = 1 WHERE id = $richiesta_id;");
-			info("approvata la richiesta id=$richiesta_id richiesta=$titolo");
-			notifica($emailDocente, $nomeCognomeDocente, $titolo, 'Approvata', '');
-			notifica($template['email_to'], $template['email_to'], $titolo, 'Approvata', '');
+			dbExec('UPDATE modulistica_richiesta SET `approvata` = 1, `messaggio` = "'.escapeString($messaggio).'", data_approvazione=now() WHERE id = '.$richiesta_id.';');
+			info("approvata la richiesta id=$richiesta_id richiesta=$titolo messaggio=$messaggio");
+			notifica($emailDocente, $nomeCognomeDocente, $titolo, ' Approvata', $messaggio);
+			notifica($template['email_to'], $template['email_to'], $titolo, ' Approvata', $messaggio);
 		} else if ($comando == 'respingi') {
-			// todo: chiede la ragione per inserirla nelle motivazioni
-			$mesaggio = "";
-			dbExec('UPDATE modulistica_richiesta SET `respinta` = 1, `messaggio_respinta` = "'.escapeString($mesaggio).'" WHERE id = '.$richiesta_id.';');
+			dbExec('UPDATE modulistica_richiesta SET `respinta` = 1, `messaggio` = "'.escapeString($messaggio).'", data_approvazione=now() WHERE id = '.$richiesta_id.';');
 			info("respinta la richiesta id=$richiesta_id richiesta=$titolo messaggio=$messaggio");
-			notifica($emailDocente, $nomeCognomeDocente, $titolo, 'Respinta', $mesaggio);
-			notifica($template['email_to'], $template['email_to'], $titolo, 'Respinta', $mesaggio);
+			notifica($emailDocente, $nomeCognomeDocente, $titolo, 'Respinta', $messaggio);
+			notifica($template['email_to'], $template['email_to'], $titolo, 'Respinta', $messaggio);
 		} else if ($comando == 'annulla') {
-			// todo: chiede la ragione per inserirla nelle motivazioni
-			$mesaggio = "";
-			dbExec('UPDATE modulistica_richiesta SET `annullata` = 1, `messaggio_respinta` = "'.escapeString($mesaggio).'" WHERE id = '.$richiesta_id.';');
+			dbExec('UPDATE modulistica_richiesta SET `annullata` = 1, `messaggio` = "'.escapeString($messaggio).'", data_approvazione=now() WHERE id = '.$richiesta_id.';');
 			info("annullata la richiesta id=$richiesta_id richiesta=$titolo messaggio=$messaggio");
-			notifica($emailDocente, $nomeCognomeDocente, $titolo, 'Annullata', $mesaggio);
-			notifica($template['email_to'], $template['email_to'], $titolo, 'Annullata', $mesaggio);
+			notifica($emailDocente, $nomeCognomeDocente, $titolo, 'Annullata', $messaggio);
+			notifica($template['email_to'], $template['email_to'], $titolo, 'Annullata', $messaggio);
 		} else {
 			warning('comando sconosciuto: comando=' . $comando . ": ignorato");
 		}
@@ -152,6 +158,8 @@ foreach(dbGetAll("SELECT * FROM modulistica_template_campo WHERE modulistica_tem
 		.btn-approva { background-color: #1CAF43; }
 		.btn-respingi { background-color: #F1003C; }
 		.btn-waiting { background-color: #f2e829; }
+		.btn-chiudi { background-color: #0ae4f0; }
+		.btn-pendente { background-color: #777777; }
 		.btn-label { padding: 5px 12px; border-radius: 8px; margin: 10px 5px; }
 	</style>
 	<title><?php echo $titolo; ?></title>
@@ -166,12 +174,12 @@ $radioUnchecked = '<div class="tick"><b><input type="checkbox" value="" style="v
 $data = '<h2 style="text-align: center">'.$titolo.'</h2>';
 
 // controlla se e' respinta o annullata e in quel caso richiede che ci sia un messaggio se non e' gia' presente
-if ($template['approva'] && ($richiesta['respinta'] || $richiesta['annullata']) && strlen($richiesta['messaggio_respinta']) == 0) {
+if ($template['approva'] && ($richiesta['respinta'] || $richiesta['annullata']) && strlen($richiesta['messaggio']) == 0) {
 	echo('<script type="text/javascript" src="../common/jquery-3.3.1-dist/jquery-3.3.1.min.js"></script>');
 	echo('<script type="text/javascript" src="../common/bootbox-4.4.0/js/bootbox.min.js"></script>');
 	echo('<script>');
 	echo('messaggio = prompt("Inserisci la motivazione");');
-	echo('$.post("../common/recordUpdate.php", {table: "modulistica_richiesta", id: '.$richiesta_id.', nome: "messaggio_respinta", valore: messaggio}, function (data, status) {location.reload();});');
+	echo('$.post("../common/recordUpdate.php", {table: "modulistica_richiesta", id: '.$richiesta_id.', nome: "messaggio", valore: messaggio}, function (data, status) {location.reload();});');
 	echo('</script>');
 	echo("</body></html>");
 	return;
@@ -182,12 +190,11 @@ if ($template['approva']) {
 	if ($richiesta['approvata']) {
 		$data .= '<h3 style="text-align: center">La richiesta &egrave; stata <span class="btn-ar btn-approva btn-label">APPROVATA</span> in data '.$dataApprovazione.'</h3>';
 	} else if ($richiesta['respinta']) {
-		//		$data .= '<h3 style="text-align: center">La richiesta &egrave; stata <b>RESPINTA</b> in data '.$dataApprovazione.'</h3>';
 		$data .= '<h3 style="text-align: center">La richiesta &egrave; stata <span class="btn-ar btn-respingi btn-label">RESPINTA</span> in data '.$dataApprovazione.'</h3>';
-		$data .= '<h3 style="text-align: center">Motivazione: '.$richiesta['messaggio_respinta'].'</h3>';
+		$data .= '<h3 style="text-align: center">Motivazione: '.$richiesta['messaggio'].'</h3>';
 	} else if ($richiesta['annullata']) {
 		$data .= '<h3 style="text-align: center">La richiesta &egrave; stata <span class="btn-ar btn-respingi btn-label">ANNULLATA</span> in data '.$dataApprovazione.'</h3>';
-		$data .= '<h3 style="text-align: center">Motivazione: '.$richiesta['messaggio_respinta'].'</h3>';
+		$data .= '<h3 style="text-align: center">Motivazione: '.$richiesta['messaggio'].'</h3>';
 	} else {
 		$data .= '<h3 style="text-align: center">La richiesta non &egrave; ancora stata approvata</h3>';
 		if ($template['approva']) {
@@ -199,40 +206,7 @@ if ($template['approva']) {
 
 $data .= "<p>I campi del modulo sono riportati qui di seguito:</p>";
 
-$data .= '<table id="campi"><tr><th>nome</th><th>valore</th><tr>';
-
-for ($i = 0; $i < count($listaEtichette); $i++) {
-    $campo = $listaEtichette[$i];
-	if ($listaTipi[$i] == 1 || $listaTipi[$i] == 2) {
-		// per tipo 1 e 2 mette solo il valore
-		$valore = $listaValori[$i];
-	} else  if ($listaTipi[$i] == 5) {
-		// per tipo 5 (textarea) inserisce il "pre")
-		$valore = '<span  style="white-space: pre-wrap;">' . $listaValori[$i] . '</span>';
-	} else  if ($listaTipi[$i] == 3 || $listaTipi[$i] == 4) {
-		// per 3 e 4 la stringa rappresenta le posizioni in cui i checkbox o radio sono settati e i testi vanno presi da lista valori del db
-		// la trasforma in una lista di stringhe esplodendo i :: come separatori
-		$listaBoxChecked = array_map('intval', explode('::', $listaValori[$i]));
-
-		$risultato = '';
-		// prende tutte le diciture dei box
-		$localiValoriSelezionabili = explode('::', $listaValoriSelezionabili[$i]);
-		for ($j = 0; $j < count($localiValoriSelezionabili); $j++) {
-			$valoreSelezionabile = $localiValoriSelezionabili[$j];
-
-			// controlla se questo deve essere marcato
-			if (in_array($j, $listaBoxChecked)) {
-				$risultato = $risultato . $chekboxChecked . $valoreSelezionabile . '</div><br/>';
-			} else {
-				$risultato = $risultato . $chekboxUnchecked . $valoreSelezionabile . '</div><br/>';
-			}
-		}
-
-		$valore = $risultato;
-	}
-    $data .= '<tr><td class="col1">'.$campo.'</td><td class="col2">'.$valore.'</td></tr>';
-}
-$data .= '</table id="campi"></p>';
+$data .= produciTabella($listaEtichette, $listaValori, $listaTipi, $listaValoriSelezionabili);
 
 $data .= "</body></html>";